refactor: moved cgn saga into separate file
